+++ src/Shakebook/Sitemap.hs
@@ -1,3 +1,12 @@
+{- |
+   Module     : Shakebook.Sitemap
+   Copyright  : Copyright (C) 2020 Daniel Firth
+   Maintainer : Daniel Firth <dan.firth@homotopic.tech
+   License    : MIT
+   Stability  : experimental
+
+Utilities from "Web.Sitemap.Gen" lifted to `MonadAction` and `FileLike`.
+-}
 module Shakebook.Sitemap where
 
 import Data.Aeson